Add 'See also' links to examples page.
Adds a new table `link` that keeps track of links between bindings.
Currently the only link is "See also", but this in future be extended to
keep track of different types of links (such as 'deprecated by' or
'alias for').

The entries for the new table are created via a list in `seed.janet`.